Refactor store to use utility functions
