+++ package.json
@@ -15,6 +15,7 @@
         "@radix-ui/react-label": "^2.1.6",
         "@radix-ui/react-slider": "^1.3.4",
         "@radix-ui/react-slot": "^1.2.2",
+        "bitaboom": "^1.1.1",
         "class-variance-authority": "^0.7.1",
         "clsx": "^2.1.1",
         "lucide-react": "^0.510.0",
