+++ src/components/toolbar.tsx
@@ -9,16 +9,10 @@ import { FormatDialog } from './format-dialog';
 import { Button } from './ui/button';
 
 export default function Toolbar() {
-    const isInitialized = useTranscriptStore((state) => state.isInitialized);
     const mergeSegments = useTranscriptStore((state) => state.mergeSegments);
     const splitSegment = useTranscriptStore((state) => state.splitSegment);
     const selectedSegments = useTranscriptStore((state) => state.selectedSegments);
     const selectedToken = useTranscriptStore((state) => state.selectedToken);
-
-    if (!isInitialized) {
-        return null;
-    }
-
     const sortedSegments = selectedSegments.toSorted((a, b) => a.start - b.start);
 
     return (
